API Documentation

SymEngine

SymEngine.AsciiArt — Type

show symengine logo

LinearAlgebra.lu — Method

LU decomposition for CDenseMatrix, dense matrices of symbolic values

Also: lufact(a, Val{:false}) for non-pivoting lu factorization

SymEngine.N — Method

N(a)
N(Val(:Type), b::Basic)

Convert a SymEngine numeric value into a Julian number

SymEngine.coeff! — Method

coeff!(a, b, x, n)

Return coefficient of x^n term, x a symbol

SymEngine.free_symbols — Method

free_symbols(ex)

Return free symbols in an expression ex as a Set

SymEngine.function_symbols — Method

function_symbols(ex)

Return function symbols in an expression ex as a Set

SymEngine.get_args — Method

get_args(ex)

Return arguments of a function call as a vector of Basic objects

SymEngine.get_libversion — Method

get_libversion()

Get libsymengine version

SymEngine.get_name — Method

get_name(ex)

Return name of function symbol

SymEngine.get_symbol — Method

get_symbol(sym)

Helper function to lookup a symbol from libsymengine

SymEngine.get_symengine_class — Method

get_symengine_class(s)

Get SymEngine class of an object s. (e.g. 1=>:Integer, 1//2 =>:Rational, sin(x) => :Sin, ...)

SymEngine.has_symbol — Method

has_symbol(ex, x)

Does expression ex contain the symbol x

SymEngine.have_component — Method

have_component(comp::String)

Check whether libsymengine was compiled with component comp

SymEngine.is_constant — Method

is_constant(ex)

Is expression constant

SymEngine.is_symbol — Method

is_symbol(x::)

Is expression a symbol

SymEngine.lambdify — Function

lambdify(ex, vars=[]; cse=false)

Evaluates a symbol-less expression or returns a function

SymEngine.series — Function

series(ex, x, x0=0, n=6)

Series expansion to order n about point x0

SymEngine.subs — Method

subs(ex, var, val)
subs(ex, d::AbstractDict)
subs(ex, y::Tuple)
subs(ex)

Substitute values into a symbolic expression.

Examples

@vars x y
ex = x^2 + y^2
subs(ex, x, 1) # 1 + y^2
subs(ex, (x, 1)) # 1 + y^2
subs(ex, (x, 1), (y,x)) # 1 + x^2, values are substituted left to right.
subs(ex, x=>1)  # alternate to subs(x, (x,1))
subs(ex, x=>1, y=>1) # ditto

SymEngine.symbols — Method

symbols(::Symbol)

Construct symbolic value.

Examples

a = symbols(:a)
x = symbols("x")
x,y = symbols("x y")
x,y,z = symbols("x,y,z")

SymEngine.@vars — Macro

@vars x y[1:5] z()

Macro to define 1 or more variables or symbolic function

Examples

@vars x y z
@vars x[1:4]
@vars u(), x

Index

SymEngine.AsciiArt

LinearAlgebra.lu

SymEngine.N

SymEngine.coeff!

SymEngine.free_symbols

SymEngine.function_symbols

SymEngine.get_args

SymEngine.get_libversion

SymEngine.get_name

SymEngine.get_symbol

SymEngine.get_symengine_class

SymEngine.has_symbol

SymEngine.have_component

SymEngine.is_constant

SymEngine.is_symbol

SymEngine.lambdify

SymEngine.series

SymEngine.subs

SymEngine.symbols

SymEngine.@vars