API Documentation

SymEngine

SymEngine.AsciiArt — Type

show symengine logo

LinearAlgebra.lu — Method

LU decomposition for CDenseMatrix, dense matrices of symbolic values

Also: lufact(a, val{:false}) for non-pivoting lu factorization

SymEngine.N — Method

Convert a SymEngine numeric value into a Julian number

SymEngine.coeff! — Method

Return coefficient of x^n term, x a symbol

SymEngine.free_symbols — Method

Return free symbols in an expression as a Set

SymEngine.function_symbols — Method

Return function symbols in an expression as a Set

SymEngine.get_args — Method

Return arguments of a function call as a vector of Basic objects

SymEngine.get_libversion — Method

Get libsymengine version

SymEngine.get_name — Method

Return name of function symbol

SymEngine.get_symbol — Method

Helper function to lookup a symbol from libsymengine

SymEngine.get_symengine_class — Method

Get SymEngine class of an object (e.g. 1=>:Integer, 1//2 =:Rational, sin(x) => :Sin, ...

SymEngine.has_symbol — Method

Does expression contain the symbol

SymEngine.have_component — Method

Check whether libsymengine was compiled with comp

SymEngine.is_constant — Method

Is expression constant

SymEngine.is_symbol — Method

Is expression a symbol

SymEngine.lambdify — Function

lambdify

evaluates a symbolless expression or returns a function

SymEngine.series — Function

Series expansion to order n about point x0

SymEngine.subs — Method

subs

Substitute values into a symbolic expression.

Examples

@vars x y
ex = x^2 + y^2
subs(ex, x, 1) # 1 + y^2
subs(ex, (x, 1)) # 1 + y^2
subs(ex, (x, 1), (y,x)) # 1 + x^2, values are substituted left to right.
subs(ex, x=>1)  # alternate to subs(x, (x,1))
subs(ex, x=>1, y=>1) # ditto

SymEngine.symbols — Method

symbols(::Symbol) construct symbolic value

Examples:

a = symbols(:a)
x = symbols("x")
x,y = symbols("x y")
x,y,z = symbols("x,y,z")

SymEngine.@vars — Macro

@vars x y[1:5] z()

Macro to define 1 or more variables or symbolic function

Example

@vars x y z
@vars x[1:4]
@vars u(), x

Index

SymEngine.AsciiArt

LinearAlgebra.lu

SymEngine.N

SymEngine.coeff!

SymEngine.free_symbols

SymEngine.function_symbols

SymEngine.get_args

SymEngine.get_libversion

SymEngine.get_name

SymEngine.get_symbol

SymEngine.get_symengine_class

SymEngine.has_symbol

SymEngine.have_component

SymEngine.is_constant

SymEngine.is_symbol

SymEngine.lambdify

SymEngine.series

SymEngine.subs

SymEngine.symbols

SymEngine.@vars